+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## [In Development](https://github.com/ppfeufer/eve-online-killboard-widget/tree/development)
 [Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.22.5...development)
-- in development
+### Changed
+- Cache handling is now done with an own database table instead of WP transient cache. This should keep the wp_options table clean.
 
 ## [0.22.5](https://github.com/ppfeufer/eve-online-killboard-widget/releases/tag/v0.22.5) - 2018-09-10
 [Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.22.4...v0.22.5)